+++ Sources/web3swift/Web3/Web3+Protocols.swift
@@ -21,10 +21,10 @@ public protocol EventParserProtocol {
     func parseTransaction(_ transaction: EthereumTransaction) async throws -> [EventParserResultProtocol]
     func parseTransactionByHash(_ hash: Data) async throws -> [EventParserResultProtocol]
     func parseBlock(_ block: Block) async throws -> [EventParserResultProtocol]
-    func parseBlockByNumber(_ blockNumber: UInt64) async throws -> [EventParserResultProtocol]
+    func parseBlockByNumber(_ blockNumber: UInt) async throws -> [EventParserResultProtocol]
     func parseTransactionPromise(_ transaction: EthereumTransaction) async throws -> [EventParserResultProtocol]
     func parseTransactionByHashPromise(_ hash: Data) async throws -> [EventParserResultProtocol]
-    func parseBlockByNumberPromise(_ blockNumber: UInt64) async throws -> [EventParserResultProtocol]
+    func parseBlockByNumberPromise(_ blockNumber: UInt) async throws -> [EventParserResultProtocol]
     func parseBlockPromise(_ block: Block) async throws -> [EventParserResultProtocol]
 }
